% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Chinook.R
\name{Format_Chinook}
\alias{Format_Chinook}
\title{Format Chinook Salmon screw trap data into capture-mark-recapture data needed for Bayesian models}
\usage{
Format_Chinook(
  RST_ops_obs_data,
  strata = 7,
  smolt.date = "07-01",
  parr.presmolt.date = "09-01",
  species,
  trap.name
)
}
\arguments{
\item{RST_ops_obs_data}{merged trap operation and observation data for a single species}

\item{strata}{length of desired strata in days (e.g. 1 week -> 7 days)}

\item{smolt.date}{"MM-DD" smolt/parr life stage classification date needed for formatting. Date is the beginning of the parr life stage.}

\item{parr.presmolt.date}{"MM-DD" parr/presmolt classification date needed for formatting. Date is the beginning of the smolt life stage.}

\item{species}{character string needed to subset data (should be identical to the species name in the RST_ops_obs_data set)}

\item{trap.name}{character string used for titles and descriptions of reports}
}
\description{
This function reconfigures Chinook Salmon screw trap data into the appropriate capture-mark-recapture format needed for the hierarchical Bayesian models.

Chinook "TU" dispositioned fish are marked individuals released upstream (n) of the screw trap on for the prior day. "RE RC" dispositioned Chinook are recaptured individuals (m) at the RST on a given date.
The sum of "TU", "TD", "NTT", "NTD", "NTR" & "NTS" dispositioned fish are captured unmarked individuals (u) at the RST on a given date.

Fry and NTS disposition individuals captured between the earliest trapping date to the specified smolt cut-off are broken into "YOYnraw", "YOYm", and "YOYu". Precocial Chinook, designated with precocial disposition, are excluded from the formatting. Yearling fish migrating after the specified smolt date are excluded from the formatting.

Years are standardized in that each year begins at the earliest trapping date and ends at the latest trapping date the RST was ever in operation since RST installation.
}
