% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Steelhead.R
\name{Format_Steelhead}
\alias{Format_Steelhead}
\title{Format Chinook Salmon screw trap data into capture-mark-recapture data needed for Bayesian models}
\usage{
Format_Steelhead(
  RST_ops_obs_data,
  strata = 7,
  smolt.date = "06-01",
  flength.cut = 80,
  species,
  trap.name
)
}
\arguments{
\item{RST_ops_obs_data}{merged trap operation and observation data for a single species}

\item{strata}{length of desired strata in days (e.g. 1 week -> 7 days)}

\item{smolt.date}{"MM-DD" smolt classification date needed for formatting}

\item{flength.cut}{lower fork length cut off for steelhead to be included in the analysis}

\item{species}{character string needed to subset data (should be identical to the species name in the RST_ops_obs_data set)}

\item{trap.name}{character string used for titles and descriptions of reports}
}
\description{
This function reconfigures Chinook Salmon screw trap data into the appropriate capture-mark-recapture format needed for the hierarchical Bayesian models.

Steelhead  "TU" dispositioned fish are marked individuals released upstream (n) of the rotary screw trap (RST) on a the prior date. "RE RC" dispositioned Steelhead are recaptured individuals (m) at the RST on a given date.
The sum of "TU", "TD", "NTT", "NTD", and "NTR" dispositioned fish are captured unmarked individuals (u) at the RST on a given date.

Individuals smaller than the fork length cut-off (flength.cut) are excluded from the formatting. Precocial Steelhead, designated with precocial disposition, are excluded from the formatting.

Years are standardized in that each year begins at the earliest trapping date and ends at the latest trapping date the RST was ever in operation since RST installation.
}
