% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatted_Data_Example.R
\docType{data}
\name{Formatted_Data_Example}
\alias{Formatted_Data_Example}
\title{Example rotary screw trap data.}
\format{
Data is formatted and ready for model functions. Normally read in as a .CSV file
}
\usage{
Formatted_Data_Example
}
\description{
This is an example data set that can be read into the model functions
}
\keyword{Screw}
\keyword{dataset}
\keyword{example}
\keyword{trap}
