% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_RST_Obs_Ops.R
\name{Merge_RST_Obs_Ops}
\alias{Merge_RST_Obs_Ops}
\title{Merge screw trap fish observation and trap operation data}
\usage{
Merge_RST_Obs_Ops(fish_observation, trap_operations)
}
\arguments{
\item{fish_observation}{trap fish observation data. Needs to include columns "FishDate","Acronym","NumberOfFish", "BroodYear", & "ForkLength"}

\item{trap_operations}{trap operation data. Needs to include columns "StartDate" and "Operation"}
}
\description{
This function merges raw IDFG screw trap fish observation and trap operation data into one file that can then be used with the
`Format_Chinook()` or `Format_Steelhead()` functions which summarizes the trap data into the capture-mark-recapture format
needed for the Bayesian models.
}
