% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spline_Calendar.R
\name{Spline_Calendar}
\alias{Spline_Calendar}
\title{Penalized-spline time stratified Bayesian estimator (calendar summary)}
\usage{
Spline_Calendar(
  data,
  effort.cor = FALSE,
  sel.years = currentyear,
  smolt.juv.date = "06-01",
  species = "",
  trap.name = "",
  den.plot = TRUE,
  trace.plot = TRUE,
  strata.length = s.length,
  burnin = 200000,
  chains = 3,
  iterations = 400000,
  thin = 200,
  boot = 5000,
  model.params = c("p", "U", "etaP1", "etaU1", "sigmaU", "sigmaP")
)
}
\arguments{
\item{data}{capture-mark-recapture data frame}

\item{effort.cor}{expands the number of unmarked fish captured by a sampling effort}

\item{sel.years}{selected year(s) to compute abundance estimates}

\item{smolt.juv.date}{"MM-DD" date to partition smolt life stage}

\item{species}{character string used for titles and descriptions of reports}

\item{trap.name}{character string used for titles and descriptions of reports}

\item{den.plot}{return density plots of MCMC chains}

\item{trace.plot}{return trace plots of MCMC chains}

\item{strata.length}{number of days in strata}

\item{burnin}{number of initial MCMC chain iterations to be discarded}

\item{chains}{number of MCMC chains (>1)}

\item{iterations}{number of MCMC iterations per chain}

\item{thin}{thin rate for MCMC chains}

\item{boot}{number of boot strap iterations to calculate yearly and life stage abundance estimates}

\item{model.params}{parameters to be returned from MCMC simulation}

\item{strata.op.min}{minimum number of years data need to have been collected in a stratum to be included in summary}
}
\description{
This function implements Bayesian methods to fit a Penalized-spline (p-spline) through U and a hierarchical structure for p between strata within a year.
The p-spline model was developed by Bonner & Schwarz (2011) and the R script to implement the model utilizes the TimeStratPetersenDiagError_fit()function
found in the BTSPAS packages.
This function was intended to be used to summarize juvenile steelhead abundances by calendar year.
}
